+++ apriltag_pose.c
@@ -2,9 +2,7 @@
 #include <stdio.h>
 
 #include "apriltag_pose.h"
-#include "apriltag_math.h"
 #include "common/homography.h"
-#include "common/image_u8x3.h"
 #include "common/diagnostic.h"
