+++ common/homography.c
@@ -26,7 +26,6 @@ either expressed or implied, of the Regents of The University of Michigan.
 */
 
 #include <math.h>
-#include <stdio.h>
 
 #include "common/matd.h"
 #include "common/zarray.h"
