+++ common/workerpool.c
@@ -30,7 +30,6 @@ either expressed or implied, of the Regents of The University of Michigan.
 #include <sched.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include <inttypes.h>
 #ifdef _WIN32
 #include <windows.h>
 #else
@@ -38,9 +37,6 @@ either expressed or implied, of the Regents of The University of Michigan.
 #endif
 
 #include "common/workerpool.h"
-#include "common/timeprofile.h"
-#include "common/math_util.h"
-#include "common/string_util.h"
 #include "common/diagnostic.h"
 
 struct workerpool {
