+++ common/zarray.c
@@ -25,8 +25,6 @@ of the authors and should not be interpreted as representing official policies,
 either expressed or implied, of the Regents of The University of Michigan.
 */
 
-#include <stdio.h>
-#include <stdlib.h>
 #include <string.h>
 
 #include "common/zarray.h"
